Converted polymers.building into a package, split up functionality among child modules
